+++ R/add_stat_label.R
@@ -87,9 +87,7 @@
 add_stat_label <- function(x, location = NULL, label = NULL) {
   updated_call_list <- c(x$call_list, list(add_stat_label = match.call()))
   # checking inputs ------------------------------------------------------------
-  if (!(inherits(x, "tbl_summary") | inherits(x, "tbl_svysummary"))) {
-    stop("`x=` must be class `tbl_summary` or `tbl_svysummary`", call. = FALSE)
-  }
+  .assert_class(x, c("tbl_summary", "tbl_svysummary"))
 
   # if `add_stat_label()` already run, return unmodified -----------------------
   if ("add_stat_label" %in% names(x$call_list)) {